✨ Full migrate to Chakra UI

+++ pages/deleteUser.jsx
@@ -1,10 +1,16 @@
 import { gql, useMutation } from "@apollo/client";
 import { useForm } from "react-hook-form";
-import { toast } from "react-toastify";
-import useUser from "../hooks/useUser";
-import { useRouter } from "next/router";
 import { purgeAllUsers } from "../apollo";
 import AdminOnlyPage from "../components/ProtectedPages/AdminOnlyPage";
+import {
+  Box,
+  Button,
+  FormControl,
+  FormErrorMessage,
+  Heading,
+  Input,
+  useToast,
+} from "@chakra-ui/react";
 
 const DELETE_USER = gql`
   mutation removeUser($username: String!) {
@@ -16,44 +22,61 @@ const DELETE_USER = gql`
 `;
 
 function DeleteUser() {
-  const { data, loading } = useUser();
-  const router = useRouter();
+  const toast = useToast();
   const { register, handleSubmit, setValue, formState } = useForm();
-  const [deleteUserMutation] = useMutation(DELETE_USER);
+  const [deleteUserMutation] = useMutation(DELETE_USER, {
+    onCompleted: async (data) => {
+      if (data.removeUser.success) {
+        setValue("username", "");
+        toast({
+          title: "계정이 삭제되었습니다!",
+          status: "success",
+        });
+        await purgeAllUsers();
+      } else {
+        toast({
+          title: data.removeUser.error,
+          status: "error",
+          duration: 1500,
+        });
+      }
+    },
+  });
 
   const onValid = async ({ username }) => {
-    if (!loading) {
-      if (data?.me?.type !== "Admin") {
-        router.push("/");
-      }
-    }
     await deleteUserMutation({
       variables: { username },
     });
-    await purgeAllUsers();
-    setValue("username", "");
-    toast.success("계정 삭제 완료!");
   };
 
   return (
     <AdminOnlyPage>
-      <div className="flex flex-col justify-center items-center w-full h-screen">
-        <h1 className="text-3xl font-bold mb-3">학생 제거</h1>
-        <form
-          onSubmit={handleSubmit(onValid)}
-          className="flex flex-col w-96 justify-center items-center"
-        >
-          <input
-            className="input"
+      <Box
+        w="full"
+        h="100vh"
+        display="flex"
+        justifyContent="center"
+        alignItems="center"
+        flexDirection="column"
+      >
+        <Heading mb={4} fontSize="3xl">
+          학생 제거
+        </Heading>
+        <FormControl as="form" onSubmit={handleSubmit(onValid)} maxW="sm">
+          <Input
             placeholder="이름"
             {...register("username", { required: true })}
           />
           {formState?.errors?.username?.message && (
-            <p>{formState?.errors?.username?.message}</p>
+            <FormErrorMessage>
+              {formState?.errors?.username?.message}
+            </FormErrorMessage>
           )}
-          <input type="submit" className="blueButton" value="학생 제거" />
-        </form>
-      </div>
+          <Button type="submit" w="full" mt={4}>
+            학생 제거
+          </Button>
+        </FormControl>
+      </Box>
     </AdminOnlyPage>
   );
 }
